chore: moving categoryitem from components to pages
