+++ src/App.jsx
@@ -1,9 +1,11 @@
-import style from './App.module.scss'
+// import style from './App.module.scss'
+import CategoryItem from './pages/CategoryItem.jsx'
 function App() {
 
   return (
     <>
       <div>
+        <CategoryItem/>
       </div>
     </>
   )
