+++ src/sass/pages/categoryItem.module.scss
@@ -1,4 +1,4 @@
-.categoryItemHeader{
+.categoryItemPage{
     font-family: 'Exo 2', sans-serif;
     
     &TopNav{
@@ -18,13 +18,12 @@
         }
     }
 
-    &Page{
+    &Container{
         display: flex;
         width: 100%;
         padding: 4%;
-
     }
-
+        
     &Image{
         width: 50%;
     }
@@ -38,34 +37,22 @@
         width: 100%;
         padding: 1% 10%;
 
-        h1{
+        &Header{
             font-size: 1.7rem;
             letter-spacing: 0.04em;
             font-weight: 550;
         }
-
-        &Price{
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
-
+        
         &Para{
             opacity: 0.7;
             line-height: 20px;
             font-family: $fontPara;
         }
 
-        &AmountBox{
+        &Price{
             display: flex;
-            gap: 25px;
-            justify-content: space-around;
+            justify-content: space-between;
             align-items: center;
-            border: 1px solid $primary;
-            width: 120px;
-            height: 45px;
-            margin-bottom: 20px;  
-            
         }
         
         &ColorImg {
@@ -76,11 +63,22 @@
             gap: 10px;
             margin-right: 10px;
         }
-
+        
         &ColorImg:hover{
             outline: 1px solid $primary;
         }
 
+        &AmountBox{
+            display: flex;
+            gap: 25px;
+            justify-content: space-around;
+            align-items: center;
+            border: 1px solid $primary;
+            width: 120px;
+            height: 45px;
+            margin-bottom: 20px;    
+        }
+        
         &Button button{
             padding: 4%;
             background-color: $accent;
@@ -88,7 +86,7 @@
             border: none;
             cursor: pointer;
         }
-
+        
         &Button button:hover{
             background-color: $accent;
             color: $secondary;
@@ -97,8 +95,8 @@
 }
 
 @media screen and (max-width:900px) {
-    .categoryItemHeader{
-        &Page{
+    .categoryItemPage{
+        &Container{
             display: block;
             width: 100%;
             padding: 4%;
